Refactor view related submodules.
* Move `neurom._compat` to `neurom.view._compat`.
*
