+++ neurom/view/tests/test_compat.py
@@ -29,15 +29,15 @@
 import os
 import numpy as np
 from nose import tools as nt
+from neurom.view import _compat
 from neurom.point_neurite.io.utils import load_neuron
 from neurom.core.types import NeuriteType
 from neurom import fst
-from neurom import _compat
 from neurom.analysis.morphmath import segment_radius as segrad
 
 
 _PWD = os.path.dirname(os.path.abspath(__file__))
-DATA_PATH = os.path.join(_PWD, '../../test_data/h5/v1/Neuron.h5')
+DATA_PATH = os.path.join(_PWD, '../../../test_data/h5/v1/Neuron.h5')
 
 NRN0 = load_neuron(DATA_PATH)
 NRN1 = fst.load_neuron(DATA_PATH)
